feat(9pfs): Separate the monitor rootfs from the container's rootfs
In order to spawn the monitor process inside the new root, we need to
mount a few extra files from the host. Therefore, the new rootfs will
contain extra files which are not needed by the guest at all. For that
reason, when we want to share the rootfs between the container and the
guest, we separate the monitor's rootfs from the container's one.
Furthermore, we bind mount the container's rootfs inside the monitor's
rootfs and share that directory with the guest.

PR: https://github.com/urunc-dev/urunc/pull/194
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>