feat(9pfs): Change annotation of useDMBlock to mountRootfs
By adding support for shared-fs between the guest and the host, we
create another way to mount the container's rootfs as the rootfs of the
guest. However, shared-fs and using the devmapper snapshot as a block
device can not work on the same time right now. Hence, we update the
annotations and change the useDMBlock to mountRootfs.

The new annotation instructs urunc to mount the rootfs of the container
as the rootfs of the guest. The choice between using the container's
rootfs as a block device or shared-fs depends on the setup of the
container (use of devmapper as a snapshotter) and the support for block
devices/shared-fs by the guest.

In particular, if the mountRootfs annotation is set and the container is
spawned with devmapper as a snapshotter, then urunc will check if the
guest supports block devices. If it does, then it will use the
container's snapshot as the block image for the guest's rootfs. If the
guest does not support block devices, then urunc will check if the guest
supports 9pfs for shared-fs. If it does support block devices or we do
not use devmapper as snapshotter, the urunc will use 9pfs to
mount the container's rootfs as the rootfs of the guest. In case, both
block devices and shared-fs are supported, then block device will be used.

If mountRootfs is not set, urunc will not mount the rootfs of the
container at all.

Also, we remove the option of setting the useDMBlock as an environment
variable, since we plan to add configuration in urunc. With the
configuration users will be able to enable/disable shared-fs or choose
how to mount the rootfs when both (block device and shared-fs) are
supported from the guest.

PR: https://github.com/urunc-dev/urunc/pull/194
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>