fix(qemu): Proper handling of Qemu mounts for non-existent files
In urunc-deploy and some Linux distributions, Qemu does not use a
seabios directory. Furthermore, in some distributios (e.g. Ubuntu or
rpi5) the /lib64 directory does not exist. Therefore, if we do not
find /lib64 and seabios under /usr/local/share or /shar/share we
should ignore them and move on with the rest of the files.

This commit fixes this, by checking the error returned from
fileFromHost.

PR: https://github.com/urunc-dev/urunc/pull/195
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>