refactor(rumprun): Refactor json construction
Rumprun accepts a json with various configuration options (e.g.
net/block) when it boots. However, it does not follow the correct json
syntax. Therefore, we can not use a typical marshal function to create
the respective json.

This patch creates a "json" string as Rumprun expects it and can read
it. Furthermore, it adds support for passing environment variables in
rumprun and handles properly the cases where block or network is not
used for the unikernel.

PR: https://github.com/urunc-dev/urunc/pull/224
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>