feat(virtiofs): Add support for virtiofs
Add support for sharing the container's rootfs with the guest's rootfs
through virtio-fs. To do that, we bind mount the virtiofsd binary, which
should be located at /usr/libexec/virtiofsd, inside the monitor rootfs
and prepare the container rootfs to be the guest's rootfs.

Currently, only Linux over Qemu is able to benefit from this feature.
However, due to the benefits of virtiofs over 9pfs, it will be the
default shared-fs option for Linux over Qemu.

In case virtiofs fails, urunc will fall back to 9pfs.

For testing, we replace the test for Linux over Qemu with 9pfs with a
new test for Linux over Qemu but using virtiofs.

PR: https://github.com/urunc-dev/urunc/pull/288
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>