fix(monitors): Let guests handle block args instead of monitors
The guest can define the block cli args of the monitor, since the VMM
interface calls the MonitorBlockCli function.  However, this function
was not compatible with Firecracker, which uses a json configuration
file to define the VM arguments.

This commit partially fixes this by updating the return value of
MonitorBlockCli, in order to use a new struct that contains the ID and
the path to the block device or the execat cli argumnets. Therefore, in
the case of Firecracker and SOlo5 guests cna make use of the ID and path
fields, while for Qemu, they can opt for defining the exact cli to
handle the block device.

Furthermore, with the above change, there is no reason for the monitor
struct to hold information about blocks. Instead the guests should be
responsible to handle how they expect the guests to appear inside the
sandbox. Therefore, decouple the block handling from the monitor and let
the guests define the cli or any other kind of args for block devices.

PR: https://github.com/urunc-dev/urunc/pull/329
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>