feat(Linux): Pass information for block devices to urunit
Using the following config this commit attaches multiple block images to
Linux guests and instructs urunit how to mount them inside the Linux
guest.

The configuration for blocks:

```
UBS
ID: <serial_id>
MP: <mountpoint>
...
UBE
```

THe serial ID is used to associate a blockd evice with a specifc
reference. Both Qemu and Firecracker can set the serial ID of a virtion
block device and hence the guest will be able to read that serial ID.
The mountpoint is filled with the same path as the one defined in the
mount struct of the container's configuration.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>

sq

PR: https://github.com/urunc-dev/urunc/pull/329
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>