feat(update initrd): Copy files from mount struct inside initrd
The Linux kernel is able to properly read and unpack multiple initrds
that are concatenated in a single file. This commit takes advantage of
this feature and in the case the Linux guest is using initrd, it appends
the files from the mount struct in the container's spec inside the given
initrd.

PR: https://github.com/urunc-dev/urunc/pull/293
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>