refactor(Exec): Improve the logic that chooses the guest rootfs
Simplify, make more extensible, testable and improve the logic that
chooses the rootfs of the guest. There are are 4 possible scenarios:
1. Using initrd
2. Using a block based rootfs from an image inside the container image
3. Using the container rootfs as a rootfs for the guest
  a. Either as a block device
  b. Or as a shared-fs (virtiofs, 9pfs)
4. No rootfs.

THe decision is made based on the annotations of the container image and
the support of guest, monitor and host.

Furthermore the new design will allow to take into consideration more
parameters, such as a user configuration for urunc.

PR: https://github.com/urunc-dev/urunc/pull/306
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>