fix(linux): Set console to ttyS0 for aarch64 in FC
We noticed that Linux kernel over Firecracker in aarch64 requires to set
the console in ttyS0 rather than ttyAMA0. Therefore, this commit
decides on the console type based on the monitor and the host
architecture. Therefore:

- For FC: always ttyS0
- For Qemu: ttyS0 in amd64 and ttyAMA0 in aarch64

However, we need to further investigate this and check when and using
which arguments we should choose the console type.

PR: https://github.com/urunc-dev/urunc/pull/328
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>