refactor(hooks): Improve error reporting and merge code
Report the error of a failed hook execution in both stderr of urunc and
in the logs. Furthermore, include the stdout and stderr of a the hook
execution to help users find the issue easier.

Moreover, refactor the hook execution code by merging the common code
between the two execution modes for the hooks (sequential/concurrent).

PR: https://github.com/urunc-dev/urunc/pull/335
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>