refactor(delete): Simplify directories removal
Refactor the cod eof thr Delete function in unikontainers to simplify
the process of rmeoving the directories that urunc has created.
Furthermore, remove the path to the monitor binary in the new
container's rootfs, covering also the case that the binary is not placed
under one of the removed directories.

PR: https://github.com/urunc-dev/urunc/pull/337
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>