fix(firecracker): do not add network config if no net device exists
Previously, the Firecracker configuration always included a network
interface entry, regardless of whether a network device was actually
assigned to the container. This caused Firecracker to attempt to open
a TUN/TAP device.

When running urunc with a non-root user (e.g., via `ctr run --user`),
this operation failed with "Operation not permitted," preventing the
VM from booting even for containers that did not require networking.

This patch modifies the configuration generation to only append the
network interface section if a valid TAP device is present in the
execution arguments.

Fixes #310

PR: https://github.com/urunc-dev/urunc/pull/354
Signed-off-by: Sankalp <sankalp25103@gmail.com>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>