fix(ipc): Handle unsuccessful setup for monitor environment
In case there was an error during the monitor execution environment
setup (e.g. vmm not found, rootfs issue) then the urunc start command
was keep waiting for a message from reexec that everything was ok. This
was leading to the urunc start process hanging there forever and of
course the container was staying like that.

THis commit fixes this issue and moreover it slightly refactors the
whole IPC between the urunc instances and reexec.

PR: https://github.com/urunc-dev/urunc/pull/357
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>