fix(block): Ignore mounts that share the same mount source
To determine whether the container rootfs or a container mount is
block-based, urunc retrieves mount information from
/proc/self/mountinfo. The corresponding function (`getMountInfo`)
currently returns only the mount source, mount point, and filesystem
type. Based on this information, urunc checks whether the guest
supports the filesystem type, and if so, attaches the mount source
as a block device to the sandbox.

However, there are cases where multiple mount entries appear to share
the same source. An exmaple is bind mounts where the mount source of a
bind mount appears to be the underlying device of the source path.

To handle this, this commit extends `getMountInfo` with an additional
check. It collects the mount sources of all non-special filesystems into
a map. The source of the mount corresponding to the target path is
intentionally excluded from this map. If at the end of the searching the
source of the target mount does not exist in the map, it is considered
safe to attach as a block device. If it does exist, it is not safe to
use, since the same source is mounted elsewhere or it is not a
block-based mount.

PR: https://github.com/urunc-dev/urunc/pull/572
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>